all new methods inside descheduler.go
